requirements.txt
pandas
selenium
gspread
oauth2client
df2gspread
webdriver-manager
requests
beautifulsoup4
python-dotenv
pyarrow 